Capture only keywords that require translation from org to jekyll.
Rename org-to-yaml-metadata to org-to-jekyll-metadata as this function
is intended to translate org keywords to their jekyll equivalents.  The
translation table is locally scoped as it is only useful within
org-to-jekyll-metadata. While there is only one key in the translation
table, it's important to make it simple to add more if needed.