Make article publication consistent with org2jekyll user config
Prior to this commit, the org-mode file (with the jekyll yaml header) was
generated alongside the original article. When the article was nested deeply,
this resulted in article not published in the right output folder.

This commit fixes that behavior by generating temporarily the org-mode article
at the org2jekyll-source-directory.