Respect org-mode option regarding toc generation
This rewrites how org2jekyll works, relying more on org-mode's publish
mechanism to do its work.

Prior to this commit, org2jekyll manipulated temporary files to prepare jekyll
yaml header as export html black. This worked most of the time except for
example for the toc:t option which was not respected (#58).

This still does the temporary manipulation except now, we let org-mode publish
correctly the toc (and all options for that matters). Then, through a dedicated
hook, the yaml header is as the last step prepended to the html published file.

Internally, this uses emacs' hook mechanism with the dedicated org-publish hook
`org-publish-after-publishing-hook`.

When activating org2jekyll mode, this installs the hook in
'org-publish-after-publishing-hook list.

When deactivating org2jekyll mode, this removes such hook.

Related #38