Remove kv dependency as we can avoid it
Reuse org-mode's `org-combine-plist` as it exists and we already depend on org.

tl; dr It's a bit much to require the kv dependency for only 2 functions.

Plus the merge function seems to have side-effects. Some data are popping up
out of previous session calls... Changing to the org-mode function, that
behavior disappeared.