Check for existing options before auto-generating metadata.
Check if any of the required options already exist in the org file when
calling org2jekyll-init-current-buffer. The user is not prompted to
enter this existing information.

Default and user specified options are added to the buffer in the order
New options are added to the buffer in the order they appear in
org2jekyll-default-template-entries and
org2jekyll-default-template-entries-extra.

Format strings are no longer used as all metadata data options were
hard coded in.